+++ backend/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+Database schema version: 47
+
+### Added
+
+- `plt-account-statements` can be exported via the REST API endpoint `/rest/export/plt-statements`.
+
 ## [2.0.25] - 2025-12-19
 
 Database schema version: 46